+++ hudi-client/hudi-flink-client/src/main/java/org/apache/hudi/io/storage/row/parquet/ParquetRowDataWriter.java
@@ -507,8 +507,7 @@ public class ParquetRowDataWriter {
       recordConsumer.startGroup();
       if (arrayData.size() > 0) {
         // align with Spark And Avro regarding the standard mode array type, see:
-        // org.apache.spark.sql.execution.datasources.parquet.ParquetSchemaConverter,
-        // org.apache.parquet.avro.AvroSchemaConverter
+        // https://github.com/apache/parquet-format/blob/master/LogicalTypes.md#lists
         //
         // <list-repetition> group <name> (LIST) {
         //   repeated group list {
