+++ hudi-common/src/main/java/org/apache/hudi/common/config/HoodieStorageConfig.java
@@ -506,6 +506,11 @@ public class HoodieStorageConfig extends HoodieConfig {
       return this;
     }
 
+    public Builder withIOFactoryClass(String ioFactoryClass) {
+      storageConfig.setValue(HOODIE_IO_FACTORY_CLASS, ioFactoryClass);
+      return this;
+    }
+
     public HoodieStorageConfig build() {
       storageConfig.setDefaults(HoodieStorageConfig.class.getName());
       return storageConfig;
