+++ hudi-flink-datasource/hudi-flink/src/main/java/org/apache/hudi/configuration/HadoopConfigurations.java
@@ -18,8 +18,6 @@
 
 package org.apache.hudi.configuration;
 
-import org.apache.hudi.common.config.HoodieStorageConfig;
-import org.apache.hudi.io.storage.row.HoodieFlinkIOFactory;
 import org.apache.hudi.util.FlinkClientUtil;
 
 import org.apache.flink.configuration.Configuration;
@@ -53,8 +51,6 @@ public class HadoopConfigurations {
     org.apache.hadoop.conf.Configuration hadoopConf = FlinkClientUtil.getHadoopConf();
     Map<String, String> options = FlinkOptions.getPropertiesWithPrefix(conf.toMap(), HADOOP_PREFIX);
     options.forEach(hadoopConf::set);
-    // Always use flink io factory
-    hadoopConf.set(HoodieStorageConfig.HOODIE_IO_FACTORY_CLASS.key(), HoodieFlinkIOFactory.class.getName());
     return hadoopConf;
   }
